feat: WebRTC APM, H.264 HW backends, plugin video HW, DTLS session split, Linux aarch64 + docs
This batch lands the work accumulated on the cleanup-onboarding branch
since b3ec6a7. It is intentionally grouped because most pieces were
authored and reviewed together; individual scopes are called out below.

Audio 3A — WebRTC Audio Processing integration
  * src/modules/audio3a: introduce webrtc_audio3a.cpp alongside the
    existing stub (audio3a.cpp) so the module can opt in to AEC3 / ANS
    / AGC / VAD via the new CMake option NIMRTC_VENDORED_WEBRTC_APM.
  * src/modules/audio3a/include/nimrtc/audio3a/audio3a.hpp / .cpp /
    plugin: split the public API and the plugin surface, expose
    process_config_t and a feature-flag so callers can run the stub
    when the upstream .a is unavailable.
  * src/modules/audio3a/tests: extend the plugin tests to cover both
    code paths.
  * src/third_party/webrtc_audio_processing/CMakeLists.txt: wrapper
    target that consumes the meson-built libwebrtc-audio-processing
    + abseil-cpp archives (built separately by build/build_webrtc_apm.*).
  * tools/fetch_webrtc_apm.py: clone + build script with mirror
    selection (freedesktop / google) and a graceful fallback path.
  * .gitignore: exclude the cloned src/ and meson build_meson/ trees —
    they are produced at build time and never vendored into git.

H.264 hardware backends
  * src/modules/h264: add hw_backends.hpp / .cpp to enumerate
    available encode/decode backends (VAAPI / DXVA / VideoToolbox /
    CUDA / NVCODEC) and dispatch by capability + preference order.
  * codec_plugin + tests: hook the HW backend into the existing
    codec plugin and exercise the selection path in tests.

Plugins — hardware video backend
  * src/plugins: introduce hw_video_backend.hpp / .cpp so the engine
    can register HW encode / decode backends through the plugin seam
    (no more direct dependency from engine to h264).
  * test_zero_copy_surface_path: new test that asserts a zero-copy
    surface round-trips through the HW backend.

DTLS session split
  * src/modules/dtls: split the monolithic dtls.cpp / dtls_prf.cpp /
    dtls_wolfssl.cpp into a session-only dtls_session.cpp that wraps
    the wolfSSL-backed implementation. The previous top-level files
    are removed.
  * tests: drop the obsolete test_dtls / test_cert_dump which have
    been superseded by the in-proc DTLS tests under tests/.
  * src/modules/dtls/include: refresh dtls.hpp and
    dtls_wolfssl_session.hpp to reflect the new layout.

Placeholder removals
  * src/modules/jb/src/placeholder.cpp, rtp/src/placeholder.cpp and
    sdp/src/placeholder.cpp are deleted — their content moved into
    the real jb_plugin / rtp_plugin / sdp_plugin sources.

Linux aarch64 cross-compile + docs
  * cmake/toolchains/aarch64-linux-gnu.cmake: new toolchain that
    wires up the existing NimRTCOptions for aarch64-linux-gnu builds.
  * cmake/NimRTCOptions.cmake / NimRTCVendored.cmake / CMakeLists.txt:
    expose Linux-aarch64 to the presets and surface a build option for
    enabling it.
  * CMakePresets.json: add a linux-aarch64 cross preset.
  * docs/plan/linux-x86_64-support.md: update with the aarch64 status.
  * docs/plan/pal-architecture.md: new — captures the
    platform-abstraction-layer split the plugin seam now relies on.

Chrome / DTLS interop driver
  * docs/plan/dtls_chrome_interop_review.md: review notes that drove
    the Chrome interop fixes.
  * interop/chrome/test_chrome_opus.html: tighten the opus
    negotiation checks.
  * interop/debug/audio_interop_diag.py + interop/docker/Dockerfile +
    interop/docker/docker-compose.yml: standalone diagnostic harness
    for the audio interop path that can run in CI via docker compose.
  * interop/README.md / signaling/signaling_proxy.py: refresh
    instructions and the signaling proxy.
  * tests/wolfssl_dtls/bridge: standalone wolfSSL DTLS responder and
    run_bridge.py used by the Chrome interop driver.

CI / dependabot
  * .github/dependabot.yml: enable weekly dependency updates for the
    GitHub Actions ecosystem used in ci.yml / interop.yml.
  * .github/workflows/ci.yml / interop.yml: pick up the new
    wolfssl_dtls/bridge + webrtc_audio_processing jobs and add the
    Linux aarch64 cross-build matrix entry.

Misc
  * src/engine + session + scheduler + video_* : plugin-seam
    plumbing, small refactors and new tests that ride along with
    the plugin + DTLS work.
  * examples/demo-p2p/demo-p2p.cpp + examples/README.md: refresh the
    demo to show the new HW backend hookup.
  * tools/run_e2e_acceptance.py + tests/wolfssl_dtls/CMakeLists.txt:
    hook the new bridge into the e2e acceptance runner.
  * CHANGELOG.md / README.md / SECURITY.md / NOTICE: update metadata
    to reflect the new modules and toolchain options.

+++ src/modules/bwe/src/bwe.cpp
@@ -181,7 +181,7 @@ struct Bwe::Impl {
 
                     if (elapsed_ms >= config_.change_interval.count()) {
                         const double increase_amount =
-                            (elapsed_ms / 1000.0) * config_.increase_bps;
+                            (static_cast<double>(elapsed_ms) / 1000.0) * config_.increase_bps;
 
                         std::uint32_t new_rate = current_bps_ +
                             static_cast<std::uint32_t>(increase_amount);
@@ -189,7 +189,7 @@ struct Bwe::Impl {
                         // Also consider per-RTT increase (CC-style boost).
                         if (feedback.rtt.count() > 0) {
                             const double rtt_factor =
-                                feedback.rtt.count() / 1000.0;  // seconds
+                                static_cast<double>(feedback.rtt.count()) / 1000.0;  // seconds
                             const double rtt_increase =
                                 current_bps_ * kIncreasePerRtt * rtt_factor;
                             new_rate += static_cast<std::uint32_t>(rtt_increase);
